qpool: move tests to the quasar nightly suite; drop the dev harness
test_pool2d.py (c-sweep, coverage matrix, unit-test special cases) +
pool_quasar_test_utils.py now live in
tests/ttnn/nightly/unit_tests/operations/experimental/quasar/, matching the
binary_ng/tm_ops convention (plain pytest, shared utils module, no arch
conftest — quasar pipelines select the directory by path; the suite also
passes on WH silicon as a cross-check).

The development harness (run_qpool.sh with its source-toggling A/B modes,
span/perf measurement tests and report scripts, the env-driven debug test)
is dropped from the PR; it remains in branch history for anyone debugging
on the sim or emulator.
